% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavethresh_fns.R
\name{AvBasis.var}
\alias{AvBasis.var}
\title{This function reconstructs variance of the mean estimate for a given wavelet basis
given a wst object}
\usage{
\method{AvBasis}{var}(wst, Ccode = TRUE, ...)
}
\description{
This function reconstructs variance of the mean estimate for a given wavelet basis
given a wst object
}

