% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavethresh_fns.R
\name{convert.var}
\alias{convert.var}
\title{This function converts variances of detail coefficients for a given wavelet basis
from wd objects to wst objects}
\usage{
\method{convert}{var}(wd, ...)
}
\description{
This function converts variances of detail coefficients for a given wavelet basis
from wd objects to wst objects
}
