% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{ff.moments}
\alias{ff.moments}
\title{return estimated E(ff(X)) and var(ff(X)) by delta method INPUT: mx =E(X) and vx = Var(X)}
\usage{
ff.moments(mx, vx)
}
\description{
return estimated E(ff(X)) and var(ff(X)) by delta method INPUT: mx =E(X) and vx = Var(X)
}
