% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{ffdashdd.moments}
\alias{ffdashdd.moments}
\title{return estimated E(ff'(X)*ff''(X)) and var(ff'(X)*ff''(X)) by delta method INPUT: mx =E(X) and vx = Var(X)}
\usage{
ffdashdd.moments(mx, vx)
}
\description{
return estimated E(ff'(X)*ff''(X)) and var(ff'(X)*ff''(X)) by delta method INPUT: mx =E(X) and vx = Var(X)
}

