% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{ffdd.moments}
\alias{ffdd.moments}
\title{Return estimated E(ff''(X)) and var(ff''(X)) by delta method, with
inputs mx = E(X) and vx = Var(X).}
\usage{
ffdd.moments(mx, vx)
}
\description{
Return estimated E(ff''(X)) and var(ff''(X)) by delta method, with
inputs mx = E(X) and vx = Var(X).
}
