% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titable_gaus.R
\name{lshift}
\alias{lshift}
\title{Shift a vector one unit to the left.}
\usage{
lshift(x)
}
\arguments{
\item{x:}{a vector}
}
\value{
a vector of the same length as that of x
}
\description{
Shift a vector one unit to the left.
}
