% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{mu.smooth}
\alias{mu.smooth}
\title{mu.smooth}
\usage{
mu.smooth(wc, data.var, basis, tsum, Wl, return.loglr, post.var, ashparam, J, n)
}
\value{
'mu.est' if posterior variances are not computed, and a list with elements 'mu.est' and 'mu.est.var' otherwise
}
\keyword{internal}

