% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titable_gaus.R
\name{reverse.gwave}
\alias{reverse.gwave}
\title{Reverse wavelet transform a set of wavelet coefficients in TItable format for Gaussian data.}
\usage{
reverse.gwave(est, lp, lq = NULL)
}
\arguments{
\item{lp:}{a J by n matrix of estimated wavelet coefficients.}

\item{lq:}{a J by n matrix of complementary wavelet coefficients.}

\item{est:}{an n-vector. Usually a constant vector with each element equal to the estimated total mean.}
}
\value{
reconstructed signal in the original data space.
}
\description{
Reverse wavelet transform a set of wavelet coefficients in TItable format for Gaussian data.
}
