% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titable_poiss.R
\name{reverse.pwave}
\alias{reverse.pwave}
\title{Reverse wavelet transform a set of probabilities in TItable
  format for Poisson data.}
\usage{
reverse.pwave(est, lp, lq = NULL)
}
\arguments{
\item{est}{An n-vector. Usually a constant vector with each
element equal to the estimated log(total rate).}

\item{lp}{A J by n matrix of probabilities.}

\item{lq}{A J by n matrix of complementary probabilities.}
}
\value{
Reconstructed signal in the original data space.
}
\description{
Reverse wavelet transform a set of probabilities in TItable
  format for Poisson data.
}
