% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titable_gaus.R
\name{rshift}
\alias{rshift}
\title{Shift a vector one unit to the right.}
\usage{
rshift(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector of the same length as that of x.
}
\description{
Shift a vector one unit to the right.
}
