% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{sd_estimate_gasser_etal}
\alias{sd_estimate_gasser_etal}
\title{Estimate homoskedastic standard deviation from nonparamatric regression}
\usage{
sd_estimate_gasser_etal(x)
}
\arguments{
\item{x}{The data.}
}
\value{
An estimate of the standard deviation
}
\description{
Estimate homoskedastic standard deviation from nonparamatric regression
}
\details{
Uses formula (3) from Brown and Levine (2007), Annals of Statistics, who attribute it to Gasser et al.
}
