% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{setAshParam.poiss}
\alias{setAshParam.poiss}
\title{Set default \code{ash} parameters.}
\usage{
setAshParam.poiss(ashparam)
}
\arguments{
\item{ashparam:}{a list of parameters to be passed to ash.}
}
\description{
Set default \code{ash} parameters.
}
\keyword{internal}
