% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{setGlmApproxParam}
\alias{setGlmApproxParam}
\title{Set default \code{glm.approx} parameters.}
\usage{
setGlmApproxParam(glm.approx.param)
}
\arguments{
\item{glm.approx.param:}{a list of parameters to be passed to glm.approx.}
}
\description{
Set default \code{glm.approx} parameters.
}
\keyword{internal}

