% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{shrink.wc}
\alias{shrink.wc}
\title{a wrapper function for code in \code{mu.smooth} and \code{var.smooth}}
\usage{
shrink.wc(wc, wc.var.sqrt, prior, pointmass, nullcheck, VB, mixsd, mixcompdist,
  gridmult, jash, df, SGD)
}
\description{
a wrapper function for code in \code{mu.smooth} and \code{var.smooth}
}
\keyword{internal}

