% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{smash.gaus}
\alias{smash.gaus}
\title{Estimate the underlying mean function from noisy Gaussian data}
\usage{
smash.gaus(x, sigma = NULL, v.est = FALSE, joint = FALSE,
  v.basis = FALSE, post.var = FALSE, filter.number = 1,
  family = "DaubExPhase", return.loglr = FALSE, jash = FALSE,
  SGD = TRUE, weight = 0.5, min.var = 1e-08, ashparam = list())
}
\arguments{
\item{filter.number, }{family: wavelet basis to be used, as in \code{wavethresh}}

\item{sigma:}{a vector of standard deviations. Can be provided if known or estimated beforehand.}

\item{v.est:}{bool, indicating if variance estimation should be performed instead.}

\item{joint:}{bool, indicating if results of mean and variance estimation should be returned together.}

\item{v.basis:}{bool, indicating if the same wavelet basis should be used for variance estimation as mean estimation. If false, defaults to Haar basis for variance estimation (this is much faster than other bases).}

\item{post.var:}{bool, indicating if the posterior variance should be returned for the mean and/or variance estiamtes.}

\item{return.loglr:}{bool, indicating if a logLR should be returned.}

\item{jash:}{indicates if the prior from method JASH should be used. This will often provide slightly better variance estimates (especially for nonsmooth variance functions), at the cost of computational efficiency. Defaults to FALSE.}

\item{SGD:}{bool, indicating if stochastic gradient descent should be used in the EM. Only applicable if jash=TRUE.}

\item{weight:}{optional parameter used in estimating overall variance. Only works for Haar basis. Defaults to 0.5. Setting this to 1 might improve variance estimation slightly}

\item{min.var:}{The minimum positive value to be set if the variance estimates are non-positive.}

\item{ashparam:}{a list of parameters to be passed to \code{ash}; default values are set by function \code{\link{setAshParam.gaus}}.}
}
\value{
\code{smash.gaus} returns the following by default:
\item{mu.res}{a list with the mean estimate, its posterior variance if \code{post.var} is TRUE, the logLR if \code{return.loglr} is TRUE, or a vector of mean estimates if neither \code{post.var} nor \code{return.loglr} are TRUE}
If \code{v.est} is TRUE, then \code{smash.gaus} returns the following:
\item{var.res}{a list with the variance estimate, its posterior variance if \code{post.var} is TRUE, or a vector of variance estimates if \code{post.var} is FALSE}
In addition, if \code{joint} is TRUE, then both \code{mu.res} and \code{var.res} are returned.
}
\description{
This function takes a data vector of length a power of 2 and performs signal denoising using wavelet decomposition and an adaptive shrinkage prior on the wavelet parameters. The data are assumed to be (mostly) independent and Gaussian, but not necessarily identically distributed.
}
\details{
We assume that the data come from the model $Y_t = \mu_t + \epsilon_t$ for t=1,...,T, where $\mu_t$ is an underlying mean, assumed to be spatially
structured (or treated as points sampled from a smooth continous function), and $\epsilon_t \sim N(0, \sigma_t)$, and are independent. 
Smash provides estimates of $\mu_t$ and $\sigma_t^2$ (and their posterior variances if desired).
}
\examples{
n=2^10
t=1:n/n
spike.f=function(x) (0.75*exp(-500*(x-0.23)^2)+1.5*exp(-2000*(x-0.33)^2)+3*exp(-8000*(x-0.47)^2)+2.25*exp(-16000*(x-0.69)^2)+0.5*exp(-32000*(x-0.83)^2))
mu.s=spike.f(t)
#Gaussian case
mu.t=(1+mu.s)/5
plot(mu.t,type='l')
var.fn=(0.0001+4*(exp(-550*(t-0.2)^2)+exp(-200*(t-0.5)^2)+exp(-950*(t-0.8)^2)))/1.35
plot(var.fn,type='l')
rsnr=sqrt(5)
sigma.t=sqrt(var.fn)/mean(sqrt(var.fn))*sd(mu.t)/rsnr^2
X.s=rnorm(n,mu.t,sigma.t)
mu.est<-smash.gaus(X.s)
plot(mu.t,type='l')
lines(mu.est,col=2)

}

