% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash.R
\name{smash}
\alias{smash}
\title{Smash a nested list into a tidy table.}
\usage{
smash(x, col.names = NULL)
}
\arguments{
\item{x}{List to smash.}

\item{Name}{for column corresponding to each level.}
}
\value{
A \code{data.table} object.
}
\description{
Smash a nested list into a tidy table.
}
