% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavethresh_fns.R
\name{threshold.haar}
\alias{threshold.haar}
\title{This is a modified threshold.wd function from package 'wavethresh', designed to perform thresholding for heteroskedastic Gaussian errors when using the Haar basis}
\usage{
\method{threshold}{haar}(vdtable, vtable, lambda.thresh, levels,
  type = "hard", policy = "universal")
}
\description{
This is a modified threshold.wd function from package 'wavethresh', designed to perform thresholding for heteroskedastic Gaussian errors when using the Haar basis
}

