% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavethresh_fns.R
\name{threshold.var}
\alias{threshold.var}
\title{This is a modified threshold.wd function from package 'wavethresh', designed to perform thresholding for heteroskedastic Gaussian errors when using non-Haar basis}
\usage{
\method{threshold}{var}(x.w, x.w.v, lambda.thresh, levels, type = "hard")
}
\description{
This is a modified threshold.wd function from package 'wavethresh', designed to perform thresholding for heteroskedastic Gaussian errors when using non-Haar basis
}

