% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titable_gaus.R
\name{titable}
\alias{titable}
\title{Produces two TI tables. One table contains the difference between adjacent pairs of data in the same resolution, and the other table contains the sum.}
\usage{
titable(sig)
}
\arguments{
\item{sig:}{a signal of length a power of 2}
}
\value{
a list of two TI tables in the form of matrices
}
\description{
Produces two TI tables. One table contains the difference between adjacent pairs of data in the same resolution, and the other table contains the sum.
}
