\name{treas}
\docType{data}
\alias{treas}
\title{Three-month treasury bill data.}
\description{
Yields of the three-month Treasury Bill from the secondary market rates,
on Fridays. The secondary market rates are annualised using a 360-day
year of bank interest and are quoted on a discount basis. The data
consist of 1735 weekly observations, from 5 January 1962 to 31 March
1995.
}
\usage{data(treas)}
\format{
  Data are stored in numeric (floating-point) vector, \code{treas}, with
  1735 entries.
}
\references{
  J. Fan and Q. Yao (1998). Efficient estimation of conditional
  variance functions in stochastic regression. \emph{Biometrika}
  \bold{85}, 645--660.
}
\examples{
  # See smashr vignette.
}
\keyword{datasets}
