% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavethresh_fns.R
\name{wd.var}
\alias{wd.var}
\title{This function performs "decomposition' of variances of detail coefficients for a given wavelet basis
in wavelet transformation}
\usage{
wd.var(data, filter.number = 10, family = "DaubLeAsymm", type = "wavelet",
  bc = "periodic", verbose = FALSE, min.scale = 0, precond = TRUE)
}
\description{
This function performs "decomposition' of variances of detail coefficients for a given wavelet basis
in wavelet transformation
}
