% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{wls.mb}
\alias{wls.mb}
\title{Returns a list with elements "coef", "se", "wrse2" if g is not
specified, or a list with elements "muhat", "semuhat", "betahat",
"sebetahat", "covmubeta", "wrse2" otherwise.}
\usage{
wls.mb(z, v, disp, indexnm, ng, df, forcebin, g = NULL, n = NULL, vv = NULL)
}
\description{
Returns a list with elements "coef", "se", "wrse2" if g is not
specified, or a list with elements "muhat", "semuhat", "betahat",
"sebetahat", "covmubeta", "wrse2" otherwise.
}
