% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{ffdashdd.moments}
\alias{ffdashdd.moments}
\title{Return estimated E(ff'(X)*ff''(X)) and var(ff'(X)*ff''(X)) by delta
method, with inputs mx = E(X) and vx = Var(X).}
\usage{
ffdashdd.moments(mx, vx)
}
\description{
Return estimated E(ff'(X)*ff''(X)) and var(ff'(X)*ff''(X)) by delta
method, with inputs mx = E(X) and vx = Var(X).
}
