% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.R
\name{ffdd.moments}
\alias{ffdd.moments}
\title{return estimated E(ff''(X)) and var(ff''(X)) by delta method INPUT: mx =E(X) and vx = Var(X)}
\usage{
ffdd.moments(mx, vx)
}
\description{
return estimated E(ff''(X)) and var(ff''(X)) by delta method INPUT: mx =E(X) and vx = Var(X)
}

