% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titable_gaus.R
\name{interleave}
\alias{interleave}
\title{Interleave two vectors}
\usage{
interleave(x, y)
}
\arguments{
\item{x:}{a vector}

\item{y:}{a vector of the same length as y}
}
\value{
a vector of length twice that of x (or y)
}
\description{
Interleave two vectors
}

