% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash.R
\name{reflect}
\alias{reflect}
\title{Reflect and extend a vector.}
\usage{
reflect(x)
}
\arguments{
\item{x}{An n-vector.}
}
\value{
A list with two list elements: \code{"x"} containing the
  extended-and-reflected signal; and \code{"idx"} containing the indices of the
  original signal.
}
\description{
Extends the vector to have length a power of 2 (if not already a power of 2) and then
  reflects it about its right end.
}
\details{
The vector x is first reflected about both its left and right ends, by (roughly) the same
amount each end, to make its length a power of 2 (if the length of x is already a power of 2 this step is skipped).
Then the resulting vector is reflected about its right end to create a vector
that is both circular (left and right ends are the same) and a power of 2.
}
