% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{smash}
\alias{smash}
\title{Estimate the underlying mean or intensity function from Gaussian or Poisson data respectively}
\usage{
smash(x, model = NULL, ...)
}
\arguments{
\item{x:}{a vector of observations}

\item{model:}{specifies the model (Gaussian or Poisson). Can be NULL, in which case the Poisson model is assumed if x
consists of integers, and the Gaussian model is assumed otherwise. One of 'gaus' or 'poiss' can also be specified to fit a specific model.}
}
\description{
This is a wrapper function for smash.gaus or smash.pois as appropriate. For details see \code{smash.gaus} and \code{smash.poiss}
}

